requirements.txt
selenium~=4.6.0
webdriver-manager~=3.8.4
behave~=1.2.6
packaging == 21.3
requests~=2.28.1
allure-behave~=2.11.1